refactor: 拆分串接 API 邏輯 將串接 API 邏輯獨立出來，並刪除部分測試的紀錄
